backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
requests==2.32.3
pyjwt==2.8.0
python-multipart==0.0.9
httpx==0.27.0
